requirements.txt
langchain_core==0.3.22
langchain_google_genai==2.0.6
langchain_openai==0.2.11
langgraph==0.2.56
python-dotenv==1.0.1
streamlit==1.40.2
